# documentation/final-project-api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
        "/users/{id}": {
            "get": {
                "tags": [
                    "user-controller"
                ],
                "operationId": "getUserById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "user-controller"
                ],
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "user-controller"
                ],
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/auth/register": {
            "post": {
                "tags": [
                    "auth-controller"
                ],
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/results/user/{userId}": {
            "get": {
                "tags": [
                    "quiz-result-controller"
                ],
                "operationId": "getResultsForUser",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/QuizResult"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "GrantedAuthority": {
                "type": "object",
                "properties": {
                    "authority": {
                        "type": "string"
                    }
                }
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "username": {
                        "type": "string",
                        "minLength": 1
                    },
                    "email": {
                        "type": "string",
                        "minLength": 1
                    },
                    "password": {
                        "type": "string",
                        "maxLength": 2147483647,
                        "minLength": 6
                    },
                    "enabled": {
                        "type": "boolean"
                    },
                    "accountNonLocked": {
                        "type": "boolean"
                    },
                    "credentialsNonExpired": {
                        "type": "boolean"
                    },
                    "accountNonExpired": {
                        "type": "boolean"
                    },
                    "authorities": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/GrantedAuthority"
                        }
                    }
                },
                "required": [
                    "email",
                    "password",
                    "username"
                ]
            },
...
            "Question": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "questionText": {
                        "type": "string",
                        "minLength": 1
                    },
                    "optionA": {
                        "type": "string",
                        "minLength": 1
                    },
                    "optionB": {
                        "type": "string",
                        "minLength": 1
                    },
                    "optionC": {
                        "type": "string"
                    },
                    "optionD": {
                        "type": "string"
                    },
                    "correctAnswer": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "correctAnswer",
                    "optionA",
                    "optionB",
                    "questionText"
                ]
            },
            "Quiz": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string",
                        "minLength": 1
                    },
                    "description": {
                        "type": "string"
                    },
                    "questions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Question"
                        }
                    }
                },
                "required": [
                    "title"
                ]
            },
            "QuizResult": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "quiz": {
                        "$ref": "#/components/schemas/Quiz"
                    },
                    "score": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "completedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
        }
    }
}